archive/ai-agents/requirements.txt
# 测试框架
pytest>=8.4.0,<9.0.0
pytest-asyncio==1.2.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0

# 环境配置
python-dotenv==1.2.1
pydantic

# Web框架
flask[async]>=2.0.0
flask-cors>=4.0.0
python-multipart>=0.0.5

# 数据库支持
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0

# 开发工具
black>=22.0.0
flake8>=4.0.0

# LangChain V1 (Stable)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.12.0
langsmith>=0.4.0,<0.5.0
